update(front): logout API

+++ molly-react/src/pages/First.js
@@ -37,8 +37,18 @@ const First = () => {
   }, []);
 
   const handleLogout = () => {
-    localStorage.clear();
-    setLogin(false);
+    fetch(`"http://localhost:8080/api/account/logout`, {
+      method : "DELETE",
+      headers : {
+        "Refresh-Token": localStorage.getItem("refreshToken"),
+      }
+      .then(res => {
+        if(res.status === 200) {
+          localStorage.clear();
+          setLogin(false);
+        }
+      })
+    })
   }
 
   return (
